V2/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.27.1
langchain
langchain-community
langchain-core
langchain-openai
openai==2.14.0
websockets==15.0.1
tiktoken==0.12.0
regex==2025.11.3
langchain-text-splitters
langchain-google-genai
kagglehub
requests
beautifulsoup4
